remove long for mask init

+++ src/mcvd/datasets/semantic_seg.py
@@ -42,7 +42,7 @@ class SegmentationMaskDataset(Dataset):
         video_num = idx // self.per_vid_data_len
         start_idx = idx % self.per_vid_data_len
         
-        all_frame_mask = torch.FloatTensor(np.load(os.path.join(self.map_idx_image_folder[video_num], 'mask.npy'))).long()[start_idx: start_idx + self.n_frames]
+        all_frame_mask = torch.FloatTensor(np.load(os.path.join(self.map_idx_image_folder[video_num], 'mask.npy')))[start_idx: start_idx + self.n_frames]
         
         all_frame_mask[all_frame_mask >= 49] = 0 # set some noise to background
         # do this padding to make image_size = 240
